# backend/http/openapi.yaml
openapi: 3.0.3
info:
  title: Engineering Insights API
  version: 1.0.0
  description: API for managing users, orgs, integrations, and engineering metrics.
servers:
  - url: http://localhost:8080/api
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    Organization:
      type: object
      properties:
        id: { type: string }
        name: { type: string }
        slug: { type: string }
        isOwner: { type: boolean }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
...
paths:
...
  /organizations:
    post:
      summary: Create an organization
      description: Creates a new organization and sets the current user as its owner.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name: { type: string }
                slug: { type: string }
              required:
                - name
                - slug
      responses:
        "201":
          description: Organization created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Organization"
      security:
        - bearerAuth: []

    get:
      summary: List user's organizations
      description: Returns a list of organizations the current user is part of, including ownership information.
      responses:
        "200":
          description: List of organizations
          content:
            application/json:
              schema:
                type: object
                properties:
                  organizations:
                    type: array
                    items:
                      $ref: "#/components/schemas/Organization"
      security:
        - bearerAuth: []
...